Moved utility functions from top of classifier.py to util.py
