+++ whistl/util.py
@@ -1,6 +1,44 @@
 '''This file contains useful functions for processing data'''
 
 
+def count_correct(output, labels):
+    '''Calculate the number of correct predictions in the given batch'''
+    # This could be more efficient with a hard sigmoid or something,
+    # Performance impact should be negligible though
+    correct = 0
+    predictions = [1 if p > .5 else 0 for p in output]
+    for y, y_hat in zip(predictions, labels):
+        if y == y_hat:
+            correct += 1
+    return correct
+
+
+def parse_map_file(map_file_path):
+    '''Create a sample: label mapping from the pickled file output by label_samples.py
+
+    Arguments
+    ---------
+    map_file_path: str or Path object
+        The path to a pickled file created by label_samples.py
+
+    Returns
+    -------
+    sample_to_label: dict
+        A string to string dict mapping sample ids to their corresponding label string.
+        E.g. {'GSM297791': 'sepsis'}
+    '''
+    sample_to_label = {}
+    label_to_sample = None
+    with open(map_file_path, 'rb') as map_file:
+        label_to_sample, _ = pickle.load(map_file)
+
+    for label in label_to_sample:
+        for sample in label_to_sample[label]:
+            sample_to_label[sample] = label
+
+    return sample_to_label
+
+
 def get_labels(df, sample_to_label, label_to_encoding):
     ''' Retrieve the labels for the given dataset
